requirements.txt
pandas
numpy
requests
scikit-learn
xgboost
fastapi
uvicorn
python-dotenv
openai
